+++ services/database/migrations/20260302082853_drop_session_table.sql
@@ -1,0 +1,17 @@
+-- migrate:up
+
+/*
+  The session API is removed, and the table no longer needed.
+*/
+
+DROP TABLE IF EXISTS "session";
+
+-- migrate:down
+
+CREATE TABLE "session" (
+    "sid" varchar NOT NULL COLLATE "default",
+    "sess" json NOT NULL,
+    "expire" timestamp(6) NOT NULL
+) WITH (OIDS=FALSE);
+ALTER TABLE "session" ADD CONSTRAINT "session_pkey" PRIMARY KEY ("sid") NOT DEFERRABLE INITIALLY IMMEDIATE;
+CREATE INDEX "IDX_session_expire" ON "session" ("expire");