+++ services/server/src/server/apiv1/verification/solc-json/solc-json.routes.ts
@@ -1,11 +1,9 @@
 import solcJsonStatelessRoutes from "./stateless/solc-json.stateless.routes";
-import solcJsonSessionRoutes from "./session/solc-json.session.routes";
 
 import { Router } from "express";
 
 const router = Router();
 
 router.use("/", solcJsonStatelessRoutes);
-router.use("/", solcJsonSessionRoutes);
 
 export default router;