+++ services/server/test/helpers/ServerFixture.ts
@@ -6,10 +6,8 @@ import config from "config";
 import { LOCAL_CHAINS } from "../../src/sourcify-chains";
 import type { StorageIdentifiers } from "../../src/server/services/storageServices/identifiers";
 import { RWStorageIdentifiers } from "../../src/server/services/storageServices/identifiers";
-import { Pool } from "pg";
+import type { Pool } from "pg";
 import type { SourcifyDatabaseService } from "../../src/server/services/storageServices/SourcifyDatabaseService";
-import genFunc from "connect-pg-simple";
-import expressSession from "express-session";
 import { SolcLocal } from "../../src/server/services/compiler/local/SolcLocal";
 import { VyperLocal } from "../../src/server/services/compiler/local/VyperLocal";
 import path from "path";
@@ -75,16 +73,6 @@ export class ServerFixture {
       ) {
         throw new Error("Not all required environment variables set");
       }
-      const PostgresqlStore = genFunc(expressSession);
-      const postgresSessionStore = new PostgresqlStore({
-        pool: new Pool({
-          host: process.env.SOURCIFY_POSTGRES_HOST,
-          database: process.env.SOURCIFY_POSTGRES_DB,
-          user: process.env.SOURCIFY_POSTGRES_USER,
-          password: process.env.SOURCIFY_POSTGRES_PASSWORD,
-          port: parseInt(process.env.SOURCIFY_POSTGRES_PORT),
-        }),
-      });
 
       const serverOptions: ServerOptions = {
         port: fixtureOptions_?.port || config.get<number>("server.port"),
@@ -100,19 +88,6 @@ export class ServerFixture {
         vyper: new VyperLocal(config.get("vyperRepo")),
         verifyDeprecated: true,
         replaceContract: true,
-        sessionOptions: {
-          secret: config.get("session.secret"),
-          name: "sourcify_vid",
-          rolling: true,
-          resave: false,
-          saveUninitialized: true,
-          cookie: {
-            maxAge: config.get("session.maxAge"),
-            secure: config.get("session.secure"),
-            sameSite: "lax",
-          },
-          store: postgresSessionStore,
-        },
         sourcifyPrivateToken: "sourcify-test-token",
         logLevel: "debug",
       };