+++ services/server/test/helpers/helpers.ts
@@ -6,7 +6,7 @@ import type {
   BytesLike,
 } from "ethers";
 import { ContractFactory, Wallet, Contract } from "ethers";
-import { assertVerificationSession, assertVerification } from "./assertions";
+import { assertVerification } from "./assertions";
 import chai, { expect } from "chai";
 import chaiHttp from "chai-http";
 import path from "path";
@@ -251,6 +251,7 @@ export function verifyAndAssertEtherscanViaApiV1(
   address: string,
   expectedStatus: VerificationStatus,
   done: Done,
+  metadataExpected: boolean = true,
 ) {
   const request = chai
     .request(serverFixture.server.app)
@@ -259,40 +260,16 @@ export function verifyAndAssertEtherscanViaApiV1(
     .field("chain", chainId);
   request.end(async (err, res) => {
     await assertVerification(
-      null,
+      serverFixture,
       err,
       res,
       done,
       address,
       chainId,
       expectedStatus,
+      metadataExpected,
     );
   });
-}
-
-export function verifyAndAssertEtherscanSession(
-  serverFixture: ServerFixture,
-  chainId: string,
-  address: string,
-  expectedStatus: VerificationStatus,
-  done: Done,
-) {
-  chai
-    .request(serverFixture.server.app)
-    .post("/session/verify/etherscan")
-    .field("address", address)
-    .field("chainId", chainId)
-    .end(async (err, res) => {
-      await assertVerificationSession(
-        null,
-        err,
-        res,
-        done,
-        address,
-        chainId,
-        expectedStatus,
-      );
-    });
 }
 
 export async function readFilesFromDirectory(dirPath: string) {