+++ packages/compilers/CHANGELOG.md
@@ -1,5 +1,13 @@
 
 
+## @ethereum-sourcify/compilers@1.0.19 - 2026-03-17
+
+- Always run solcjs in a worker (#2652)
+
+## @ethereum-sourcify/compilers@1.0.18 - 2026-03-02
+
+- update dependencies
+
 ## @ethereum-sourcify/compilers@1.0.17 - 2026-01-20
 
 - Improve log levels (#2601)