+++ services/4byte/CHANGELOG.md
@@ -1,5 +1,13 @@
 
 
+## sourcify-4byte@1.1.9 - 2026-03-17
+
+- update dependencies
+
+## sourcify-4byte@1.1.8 - 2026-03-02
+
+- update dependencies
+
 ## sourcify-4byte@1.1.7 - 2026-01-20
 
 - Update dependencies